Move `ContactsLoader` to the services folder
