+++ src/components/services/Services.js
@@ -17,7 +17,6 @@ import { setIntlLocale } from '../../locales/intl.js'
 import { selectLocale } from '../../locales/strings.js'
 import { type Action } from '../../modules/Action.js'
 import { type State } from '../../modules/ReduxTypes.js'
-import ContactsLoader from '../../modules/UI/components/ContactsLoader/contactsLoaderConnector.js'
 import { rootReducer } from '../../reducers/RootReducer.js'
 import errorAlert from '../../util/errorAlert.js'
 import loginStatusChecker from '../../util/loginStatusChecker.js'
@@ -26,6 +25,7 @@ import soundsMiddleware from '../../util/soundsMiddleware.js'
 import { ModalProvider } from '../common/ModalProvider.js'
 import { Airship } from './AirshipInstance.js'
 import { AutoLogout } from './AutoLogout.js'
+import { ContactsLoader } from './ContactsLoader.js'
 import EdgeAccountCallbackManager from './EdgeAccountCallbackManager.js'
 import EdgeContextCallbackManager from './EdgeContextCallbackManager.js'
 import EdgeWalletsCallbackManager from './EdgeWalletsCallbackManager.js'
